+++ examples/baseball/baseball_example_monte_carlo_sim.py
@@ -19,7 +19,7 @@ mph2mps = 0.44704
 def baseball_example_monte_carlo_sim():
     ## Define sim
     sim = mc.Sim(name='baseball', ndraws=ndraws, fcns=fcns, firstcaseismedian=True,
-                 seed=seed, cores=1, verbose=True, debug=True,
+                 seed=seed, singlethreaded=False, verbose=True, debug=True,
                  savecasedata=False, savesimdata=False)
 
     ## Define input variables